requirements.txt
torch
torchvision
pandas
numpy
Pillow
streamlit
scikit-learn
